Extract common bits of editable combo box behaviour
