+++ test/lib/widgets/storage/disk_device_test.rb
@@ -20,28 +20,11 @@
 require_relative "../../../test_helper"
 require "autoinstall/widgets/storage/disk_device"
 require "cwm/rspec"
+require_relative TESTS_PATH.join("support", "editable_combo_box_examples").to_s
 
 describe Y2Autoinstallation::Widgets::Storage::DiskDevice do
   subject { described_class.new }
 
   include_examples "CWM::ComboBox"
-
-  describe "#value=" do
-    let(:unknown) { "/dev/test" }
-
-    context "when an unknown value is given" do
-      it "adds the value to the list of items" do
-        subject.value = unknown
-        expect(subject.items.map(&:first)).to include(unknown)
-      end
-
-      it "sets the current value" do
-        allow(Yast::UI).to receive(:ChangeWidget)
-          .with(anything, :Items, anything)
-        expect(Yast::UI).to receive(:ChangeWidget)
-          .with(anything, :Value, unknown)
-        subject.value = unknown
-      end
-    end
-  end
+  include_examples "EditableComboBox"
 end
